Move config to separate module
